Fix transaction during scheduled importing with direct processing during the reading.
