+++ helfomat-infrastructure-typo3/src/main/java/de/helfenkannjeder/helfomat/infrastructure/typo3/Typo3OrganisationProcessor.java
@@ -20,9 +20,10 @@ import de.helfenkannjeder.helfomat.infrastructure.typo3.domain.TOrganisationType
 import de.helfenkannjeder.helfomat.infrastructure.typo3.domain.TWorkingHour;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.batch.core.configuration.annotation.JobScope;
 import org.springframework.batch.item.ItemProcessor;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.DayOfWeek;
 import java.time.LocalTime;
@@ -37,7 +38,7 @@ import java.util.stream.Collectors;
  * @author Valentin Zickner
  */
 @Component
-@JobScope
+@Transactional(propagation = Propagation.REQUIRED)
 public class Typo3OrganisationProcessor implements ItemProcessor<TOrganisation, Organisation> {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(Typo3OrganisationProcessor.class);
